requirements.txt
dj-database-url==0.4.2
dj-static==0.0.6
Django==2.0
gunicorn==19.4.1
python-decouple==3.1
pytz==2017.3
static3==0.7.0
gunicorn==19.4.1
psycopg2
django_extensions
certifi==2017.11.5
chardet==3.0.4
djangorestframework==3.7.7
flake8==3.5.0
idna==2.6
mccabe==0.6.1
pew==1.1.2
pipenv==9.0.1
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
pytz==2017.3
requests==2.18.4
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
Markdown==2.3.1
PyYAML==3.10
defusedxml==0.4.1
newrelic==2.12.0.10
pystache==0.5.3
static==1.0.2
